+++ cluster-operator/main.go
@@ -9,6 +9,7 @@ import (
 
 	clustersv1alpha1 "github.com/verrazzano/verrazzano/cluster-operator/apis/clusters/v1alpha1"
 	"github.com/verrazzano/verrazzano/cluster-operator/internal/operatorinit"
+	"github.com/verrazzano/verrazzano/pkg/constants"
 	vzlog "github.com/verrazzano/verrazzano/pkg/log"
 	"github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1beta1"
 	"go.uber.org/zap"
@@ -32,6 +33,7 @@ var (
 	runWebhooks          bool
 	runWebhookInit       bool
 	certDir              string
+	ingressHost          string
 )
 
 func init() {
@@ -59,7 +61,7 @@ func main() {
 			os.Exit(1)
 		}
 	} else {
-		err := operatorinit.StartClusterOperator(metricsAddr, enableLeaderElection, probeAddr, log, scheme)
+		err := operatorinit.StartClusterOperator(metricsAddr, enableLeaderElection, probeAddr, ingressHost, log, scheme)
 		if err != nil {
 			os.Exit(1)
 		}
@@ -78,6 +80,7 @@ func handleFlags() {
 	flag.BoolVar(&runWebhookInit, "run-webhook-init", false,
 		"Runs the webhook initialization code")
 	flag.StringVar(&certDir, "cert-dir", "/etc/certs/", "The directory containing tls.crt and tls.key.")
+	flag.StringVar(&ingressHost, "ingress-host", constants.DefaultRancherIngressHost, "The host used for Rancher API requests.")
 
 	opts := kzap.Options{}
 	opts.BindFlags(flag.CommandLine)