+++ CHANGES.md
@@ -2,8 +2,8 @@
 Component version updates:
 
 - Rancher v2.7.1
-- WebLogic Kubernetes Operator v4.0.5
-- WebLogic Monitoring Exporter v2.1.2
+- WebLogic Kubernetes Operator v4.0.6
+- WebLogic Monitoring Exporter v2.1.3
 - Jaeger v1.42.0
 
 Fixes: